+++ src/ui/view/Convert.h
@@ -5,7 +5,7 @@
 #include <optional>
 #include "Task.pb.h"
 #include "ui/step/Type.h"
-#include "utilities/TasksSortBy.h"
+#include "model/TasksSortBy.h"
 #include "ui/command/Error.h"
 #include "utilities/AliasesProtoObjects.h"
