+++ src/utilities/SortingProtoObjects.h
@@ -2,7 +2,7 @@
 
 #include <algorithm>
 #include "Task.pb.h"
-#include "TasksSortBy.h"
+#include "model/TasksSortBy.h"
 #include "AliasesProtoObjects.h"
 #include "ComparisonProtoObjects.h"
